Implemented a stub TradingFilesModel
The idea is that this class will be responsible for managing all the CSV
files. It will create/handle readers for all of them as well.

Also, it doesn't handle Drag&Drop at all, and I did NOT waste the last
hour and a half trying to make it do so...